+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All changes will be updated here.
 
+## [0.2.1] - 2022-02-16
+- Google Authenticator Library Added
+- CURD fully-written and functional
+
+
 ## [0.2.0] - 2022-02-02
 - Issue reported fixed [#4](https://github.com/SyedMuradAliShah/codeigniter-3-files-creator/issues/4)